test: 수익률 검증 테스트 코드 추가
- 당첨 금액에 따라 수익률을 올바르게 반환하는지 검증하는 테스트 코드를 추가합니다.
- 당첨 금액의 자료형을 변경합니다. 이와 관련된 파일의 자료형 또한 변경합니다.

+++ src/main/java/lotto/service/calculator/ResultCalculator.java
@@ -16,9 +16,9 @@ import lotto.strategy.SecondPlace;
 
 public class ResultCalculator {
 
-    private final EnumMap<Place, Integer> placeMap = new EnumMap<>(Place.class);
+    private final EnumMap<Place, Long> placeMap = new EnumMap<>(Place.class);
     private final Map<Integer, PlaceAuction> placeAuctionMap = new HashMap<>();
-    private final EnumMap<Place, Integer> prize = Prize.getPrize();
+    private final EnumMap<Place, Long> prize = Prize.getPrize();
 
     private ResultCalculator(List<Integer> winningResult, List<Integer> bonusResult) {
         init(bonusResult);
@@ -31,7 +31,7 @@ public class ResultCalculator {
 
     private void init(List<Integer> bonusResult) {
         for (Place place : Place.values()) {
-            placeMap.put(place, 0);
+            placeMap.put(place, 0L);
         }
         placeAuctionMap.put(6, new FirstPlace(placeMap));
         placeAuctionMap.put(5, new SecondPlace(placeMap, bonusResult));
@@ -68,8 +68,8 @@ public class ResultCalculator {
         return printResult;
     }
 
-    public Integer getPrizeMoney() {
-        int prizeMoney = 0;
+    public Long getPrizeMoney() {
+        long prizeMoney = 0;
         for (Place place : Place.values()) {
             prizeMoney += placeMap.get(place) * prize.get(place);
         }
